+++ config/world.php
@@ -24,6 +24,111 @@ return [
 	],
 	/*
 	|--------------------------------------------------------------------------
+	| Migrations.
+	|--------------------------------------------------------------------------
+	*/
+	'migrations' => [
+		'countries' => [
+			'table_name' => 'countries',
+			'optional_fields' => [
+				'phone_code' => [
+					'required' => true,
+					'type' => 'string',
+					'length' => 5,
+				],
+				'iso3' => [
+					'required' => true,
+					'type' => 'string',
+					'length' => 3,
+				],
+				'native' => [
+					'required' => false,
+					'type' => 'string',
+				],
+				'region' => [
+					'required' => true,
+					'type' => 'string',
+				],
+				'sub_region' => [
+					'required' => true,
+					'type' => 'string',
+				],
+				'latitude' => [
+					'required' => false,
+					'type' => 'string',
+				],
+				'longitude' => [
+					'required' => false,
+					'type' => 'string',
+				],
+				'emoji' => [
+					'required' => false,
+					'type' => 'string',
+				],
+				'emojiU' => [
+					'required' => false,
+					'type' => 'string',
+				],
+			],
+		],
+		'states' => [
+			'table_name' => 'states',
+			'optional_fields' => [
+				'country_code' => [
+					'required' => false,
+					'type' => 'string',
+					'length' => 3,
+				],
+				'state_code' => [
+					'required' => false,
+					'type' => 'string',
+					'length' => 3,
+				],
+				'latitude' => [
+					'required' => false,
+					'type' => 'string',
+				],
+				'longitude' => [
+					'required' => false,
+					'type' => 'string',
+				],
+			],
+		],
+		'cities' => [
+			'table_name' => 'cities',
+			'optional_fields' => [
+				'country_code' => [
+					'required' => false,
+					'type' => 'string',
+					'length' => 3,
+				],
+				'state_code' => [
+					'required' => false,
+					'type' => 'string',
+					'length' => 3,
+				],
+				'latitude' => [
+					'required' => false,
+					'type' => 'string',
+				],
+				'longitude' => [
+					'required' => false,
+					'type' => 'string',
+				],
+			],
+		],
+		'timezones' => [
+			'table_name' => 'timezones',
+		],
+		'currencies' => [
+			'table_name' => 'currencies',
+		],
+		'languages' => [
+			'table_name' => 'languages',
+		],
+	],
+	/*
+	|--------------------------------------------------------------------------
 	| Enabled seeders.
 	| The cities seeder depends on the states seeder.
 	|--------------------------------------------------------------------------