+++ src/Database/Migrations/2021_10_19_071730_create_states_table.php
@@ -13,10 +13,16 @@ class CreateStatesTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create('states', function (Blueprint $table) {
+		Schema::create(config('world.migrations.states.table_name'), function (Blueprint $table) {
 			$table->id();
 			$table->foreignId('country_id');
 			$table->string('name');
+
+			foreach (config('world.migrations.states.optional_fields') as $field => $value) {
+				if ($value['required']) {
+					$table->string($field, $value['length'] ?? null);
+				}
+			}
 		});
 	}
 
@@ -27,6 +33,6 @@ class CreateStatesTable extends Migration
 	 */
 	public function down()
 	{
-		Schema::dropIfExists('states');
+		Schema::dropIfExists(config('world.migrations.states.table_name'));
 	}
 }