+++ src/Models/Country.php
@@ -2,6 +2,7 @@
 
 namespace Nnjeim\World\Models;
 
+use Illuminate\Database\Eloquent\Concerns\GuardsAttributes;
 use Nnjeim\World\Models\Traits\CountryRelations;
 
 use Illuminate\Database\Eloquent\Model;
@@ -10,15 +11,17 @@ class Country extends Model
 {
 	use CountryRelations;
 
-	protected $fillable = [
-		'iso2',
-		'iso3',
-		'name',
-		'phone_code',
-		'region',
-		'sub_region',
-		'status',
-	];
+	protected $guarded = [];
 
 	public $timestamps = false;
+
+	/**
+	 * Get the table associated with the model.
+	 *
+	 * @return string
+	 */
+	public function getTable(): string
+	{
+		return config('world.migrations.countries.table_name', parent::getTable());
+	}
 }