+++ src/Models/Currency.php
@@ -23,4 +23,14 @@ class Currency extends Model
 	];
 
 	public $timestamps = false;
+
+	/**
+	 * Get the table associated with the model.
+	 *
+	 * @return string
+	 */
+	public function getTable(): string
+	{
+		return config('world.migrations.currencies.table_name', parent::getTable());
+	}
 }