+++ src/Models/Language.php
@@ -14,4 +14,14 @@ class Language extends Model
 	];
 
 	public $timestamps = false;
+
+	/**
+	 * Get the table associated with the model.
+	 *
+	 * @return string
+	 */
+	public function getTable(): string
+	{
+		return config('world.migrations.languages.table_name', parent::getTable());
+	}
 }