+++ src/Models/State.php
@@ -10,10 +10,17 @@ class State extends Model
 {
 	use StateRelations;
 
-	protected $fillable = [
-		'country_id',
-		'name',
-	];
+	protected $guarded = [];
 
 	public $timestamps = false;
+
+	/**
+	 * Get the table associated with the model.
+	 *
+	 * @return string
+	 */
+	public function getTable(): string
+	{
+		return config('world.migrations.states.table_name', parent::getTable());
+	}
 }