+++ src/Models/Timezone.php
@@ -16,4 +16,14 @@ class Timezone extends Model
 	];
 
 	public $timestamps = false;
+
+	/**
+	 * Get the table associated with the model.
+	 *
+	 * @return string
+	 */
+	public function getTable(): string
+	{
+		return config('world.migrations.timezones.table_name', parent::getTable());
+	}
 }